deleted registerControllerAndService
